Added tag sorting to UI
This implementation utilizes the pre-existing DatasetTagEditor.sort_tags method.

Currently sorting is hardcoded as ascending alphabetical ("123abc"), but
it should be trivial to provide a UI for the various other options
provided by the method if anyone particularly wants alternative sorting.

Tags may be sorted as a batch for all filtered tags via a new button
- Found under: Batch Edit Captions > Extras > Sort selected tags
- The "Extras" menu was created as part of this commit

There is also an option to automatically sort tags upon clicking the
"Apply changes to selected image" button when editing tags via the
"Edit Caption of Selected Image" menu
- Disabled by default
- Enabled via: "Sort caption on save" checkbox (same menu)